#42 Refactor
