+++ scripts/common.js
@@ -159,3 +159,25 @@ export const variantsClassesToBEM = (blockClasses, expectedVariantsNames, blockN
     }
   });
 };
+
+/**
+ *
+ * @param {string} blockName - block name with '-' instead of spaces
+ * @param {string} blockContent - the content that will be set as block inner HTML
+ * @param {object} options - other options like variantsClasses
+ * @returns
+ */
+export async function loadAsBlock(blockName, blockContent, options = {}) {
+  const { variantsClasses = [] } = options;
+  const blockEl = createElement('div', {
+    classes: ['block', blockName, ...variantsClasses],
+    props: { 'data-block-name': blockName },
+  });
+  const wrapperEl = createElement('div');
+  wrapperEl.append(blockEl);
+
+  blockEl.innerHTML = blockContent;
+  await loadBlocks(wrapperEl);
+
+  return blockEl;
+}
